% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvordnorm.R
\name{coef.mvordnorm}
\alias{coef.mvordnorm}
\title{Coefficients of Multivariate Models with Ordinal and Gaussian Responses}
\usage{
\method{coef}{mvordnorm}(object, ...)
}
\arguments{
\item{object}{an object of class \code{'mvordnorm'}.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
\code{coef} is a generic function which extracts the coefficients from objects of class \cr
\code{'mvordnorm'}.
}
