% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvordnorm.R
\docType{data}
\name{data_toy}
\alias{data_toy}
\title{Data set toy example}
\format{
A data frame with 1000 rows and 7 variables
}
\usage{
data("data_toy", package = "mvordnorm")
}
\description{
A simulated data set containing 8 variables (5 responses variables and 3 covariates).
\itemize{
  \item \code{z1} numeric, gaussian response variable
  \item \code{z2} numeric, gaussian response variable
  \item \code{y1} integer, ordinal response variable
  \item \code{y2} integer, ordinal response variable
  \item \code{X1} continuous covariate
  \item \code{X2} continuous covariate
  \item \code{X3} continuous covariate
}
}
