% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvordnorm.R
\name{logLik.mvordnorm}
\alias{logLik.mvordnorm}
\title{Pairwise Log-Likelihood of Multivariate Models with Ordinal and Gaussian Responses.}
\usage{
\method{logLik}{mvordnorm}(object, ...)
}
\arguments{
\item{object}{an object of class \code{'mvordnorm'}.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
\code{logLik} is a generic function which extracts the pairwise log-likelihood from objects of class \cr
\code{'mvordnorm'}.
}
