% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvordnorm.R
\name{mvordnorm}
\alias{mvordnorm}
\title{Fitting a multivariate model of mixed normal and ordinal responses}
\usage{
mvordnorm(
  formula,
  data,
  response_types = NULL,
  na.action,
  weights,
  offset = NULL,
  contrasts = NULL,
  control = mvordnorm.control(),
  ...
)
}
\arguments{
\item{formula}{A formula as in the Formula package}

\item{data}{data frame}

\item{response_types}{a (named) vector of characters. Each element of the vector is
either "gaussian" or "ordinal".}

\item{na.action}{a function which indicates what should happen when the data contain NAs.}

\item{weights}{weights which need to be constant across multiple measurements. Negative weights are not allowed.}

\item{offset}{this can be used to specify an a priori known component to be included in the linear predictor during fitting.}

\item{contrasts}{an optional list. See the \code{contrasts.arg} of \code{\link{model.matrix.default}}.}

\item{control}{list of parameters for controlling the fitting process. See \code{\link{mvordnorm.control}} for details.}

\item{...}{additional arguments.}
}
\value{
an object of class mvordnorm
}
\description{
mvordnorm is used to fit a multivariate model of mixed normal and ordinal random variables. Probit link is used for the ordinal models
}
\examples{

\dontrun{
library("mvordnorm")
data("data_toy", package = "mvordnorm")
fit <- mvordnorm("y1 + y2 + z1 + z2 ~ 0 + X1 + X2 + X3", data = data_toy,
                response_types = c("ordinal", "ordinal","gaussian", "gaussian"),
                control = mvordnorm.control(se = TRUE, solver = "CG"))
}
}
