% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvordnorm.R
\name{print.mvordnorm}
\alias{print.mvordnorm}
\title{Print Method for class mvordnorm}
\usage{
\method{print}{mvordnorm}(x, call = FALSE, ...)
}
\arguments{
\item{x}{object of class \code{'mvordnorm'}}

\item{call}{displays function call if \code{TRUE}}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Prints regression coefficients, parameters of the error structure of class mvordnorm
}
