% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvordnorm.R
\name{print.summary.mvordnorm}
\alias{print.summary.mvordnorm}
\title{Print method for summary of mvordnorm}
\usage{
\method{print}{summary.mvordnorm}(x, digits = max(1L, getOption("digits") - 3L), ...)
}
\arguments{
\item{x}{object of class \code{'summary.mvordnorm'}}

\item{digits}{number of digits in output}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Summary of thresholds, regression coefficients
and parameters of the error structure of class \code{'mvordnorm'}.
}
