% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvordnorm.R
\name{summary.mvordnorm}
\alias{summary.mvordnorm}
\title{Summary method for mvordnorm}
\usage{
\method{summary}{mvordnorm}(object, call = FALSE, ...)
}
\arguments{
\item{object}{object of class \code{'mvordnorm'}}

\item{call}{displays function call if \code{TRUE}}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Summary of thresholds, regression coefficients
and parameters of the error structure of class \code{'mvordnorm'}.
}
