% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvordnorm.R
\name{vcov.mvordnorm}
\alias{vcov.mvordnorm}
\title{vcov of Multivariate Models with Ordinal and Gaussian Responses.}
\usage{
\method{vcov}{mvordnorm}(object, ...)
}
\arguments{
\item{object}{an object of class \code{'mvordnorm'}.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
\code{vcov} is a generic function which extracts the Godambe information matrix from objects of class \cr
\code{'mvordnorm'}.
}
