% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvordnorm.R
\name{mvordnorm.control}
\alias{mvordnorm.control}
\title{Summary method for mixoglmm.}
\usage{
\method{summary}{mixoglmm}(se = TRUE, solver = "CG")
}
\arguments{
\item{se}{logical, should standard errors be calculated?}

\item{solver}{name of solver.}

\item{object}{object of class \code{'mixoglmm'}}

\item{call}{displays function call if \code{TRUE}}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Summary of thresholds, regression coefficients
and parameters of the error structure of class \code{'mixoglmm'}.

Control arguments are set.
}
