% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvordnorm.R
\name{mvordnorm.control}
\alias{mvordnorm.control}
\title{Control function for mvordnorm()}
\usage{
mvordnorm.control(se = TRUE, solver = "CG", usegrfun = TRUE)
}
\arguments{
\item{se}{logical, indicating whether standard errors should be calculated.}

\item{solver}{name of solver to be used by optimx.}
}
\description{
Control arguments are set.
}
